[#71] Refactored + Added feature code
Refactoring:
• staticfile() was refactored to allow for url_for() to be called as
follows: url_for('views.staticfile', folder='<foldername>', filename='<filename>'). This resolves the issue of this method not returning files properly and not returning files from the js folder at all. Note that the js folder was added back to the file structure.
• Imported send_file
• Refactored views.question to function as a controller where a method
and one or more arguments can be passed. This decouples the method from
a specific functionality implemented within the "controller" and allows
for a modular approach to developing new functionalities within the
"controller."

New Feature:
• Implemented the accepted answer feature in views,question.
• Modified the logic that gets a new question to account for the new
feature in processing the list of answers and the fron-end behaviour.
• Added TODOS for further improvement.